+++ packages/ui/src/components/molecules/index.ts
@@ -1,7 +1,12 @@
-import { BackButton } from './BackButton'
-import { Tabs } from './Tabs'
-import { CreateAuthor } from './CreateAuthor'
 
-export { InputGroup } from './InputGroup'
-export { Modal } from './Modal'
-export { BackButton, Tabs, CreateAuthor }
+export {
+  AccessModalContent,
+  type TAccessModalContent,
+} from './AccessModalContent'
+export { BackButton, type TBackButton } from './BackButton'
+export { CardProfile, type TCardProfile } from './CardProfile'
+export { InputGroup, type TInputGroup } from './InputGroup'
+export { Modal, type TModal } from './Modal'
+export { SelectGroup, type TSelectGroup } from './SelectGroup'
+export { Tabs, type TTabs } from './Tabs'
+export { CreateAuthor, type TCreateAuthor } from './CreateAuthor'