+++ sso_front/src/App.tsx
@@ -1,1 +1,0 @@
-export const App = () => <div>Front SSO</div>;